Split `GetShortestPath` methods to smaller files
